fix(security): TD-5 replace event.toString()/open.toString() with getSimpleName() and widen xfoSummary totalLen to long (CWE-117, D-13/SR-8)
TD-5-A: default branch in processTechnicalEvent now uses
event.getClass().getSimpleName() instead of %s.formatted(event),
eliminating the peer-controlled toString() path (CWE-117).

TD-5-B: Open branch now uses open.getClass().getSimpleName() instead
of %s.formatted(open), consistent with ADR-TD5-B (preventive fix for
latent risk if Bigbone adds fields to TechnicalEvent.Open).

TD-5-C: int totalLen widened to long totalLen in LogScrubber.xfoSummary
to eliminate theoretical int overflow for large peer-supplied lists.
Javadoc updated to document the long accumulator and int count bound.

Legacy tests onEvent_EventTechnicalOpen and onEvent_EventTechnicalUnknown
updated to assert the new (class=...) format instead of raw toString().

+++ src/main/java/de/seism0saurus/glacier/util/LogScrubber.java
@@ -203,6 +203,12 @@ public final class LogScrubber {
      * <p>null elements count toward {@code xfo-values} (slot count) and contribute 0 to
      * {@code xfo-totallen} (length).
      *
+     * <p>The {@code xfo-totallen} accumulator is a {@code long} to avoid overflow when a
+     * peer-supplied list contains many large values (theoretical: M can exceed
+     * {@code Integer.MAX_VALUE} if enough large header values are provided). The element
+     * count ({@code xfo-values}) is bounded by {@link java.util.List#size()}, which is
+     * an {@code int}. TD-5-C / ADR-TD5-C.
+     *
      * <p>Always use this method when a log statement would otherwise include the raw
      * {@code X-Frame-Options} list as an argument.
      *
@@ -214,7 +220,7 @@ public final class LogScrubber {
             return "xfo-values=0 xfo-totallen=0";
         }
         int count = values.size();
-        int totalLen = 0;
+        long totalLen = 0L;
         for (String value : values) {
             if (value != null) {
                 totalLen += value.length();
